Modificato movimento su minimappa

+++ cub3d.h
@@ -133,8 +133,8 @@ typedef struct	s_map
 
 typedef struct	s_minimap
 {
-	float	pos_x;
-	float	pos_y;
+	int	pos_x;
+	int	pos_y;
 	float	ray_x;
 	float	ray_y;
 	float	move_speed;
